Make tests run well
* Because of adding dict in url path, the key of dict must be added to
views.py function like :
  def my_function(request, dict_key=None)
Like that dict parameters are passed in arguments and argument is set to
None by default
* Thoses changes said before where done in authentication/views.py
It might be done in others apps views.
* Add missing url path in urls_m.py to make tests run well. The added
lines might be remove or at least checked. I think that it was making an
error because thoses paths are called in
asso_tn/base_transport_nantes.html

* Right now, tests run well on both urls_tn.py and urls_m.py

issue #51