Make token to be used once and tests run
* We need a persistent boolean somewhere to keep a track of whether the
auth token has been already used or not. I put this on Profile model
under the name of once_token set to False by default.
* once_token boolean is turned to True when called for auth token maker.
Then it is checked by token validator. If once_token is True, it is the
first time that the token is used and once_token is turned back to
False => user is redirected to index page as wanted.
Else if once_token is False, it means that token has already been used
and user is redirected to token invalide page as it should be.
* Test in asso_tn has been changed a little bit to include above
modifications. But it should be split in 3 parts because an error due to
the fact that token is only usable one time now.

issue #51